setup: Handle restack not being installed
This changes how setup is done to rely on a shell script which delegates
to `restack edit` if installed. Otherwise, it informs the user how to
install restack and falls back to their usual editor.

This makes for a friendlier error case when the user's dotfiles are
installed on a new system.