restack/edit: Interpret GIT_EDITOR using sh
As per [`git-var`][1], the `GIT_EDITOR` variable is meant to be
interpreted by a shell, not as a path to an executable.

So the following is possible and valid.

    export GIT_EDITOR="FOO=BAR myeditor --somearg"

This changes `restack edit` to invoke the editor using `sh -c` instead
of calling directly. This is equivalent to calling,

    sh -c "$GIT_EDITOR $1" "restack" "$filepath"

(The "restack" in there is the `$0` for the `sh -c` command.)

[1]: https://git-scm.com/docs/git-var#git-var-GITEDITOR