edit: Degrade for /tmp on different partition
On Linux, if `/tmp` is on a different partition than
the directory that holds .git,
attempts to rename with `os.Rename` will fail with:

    invalid cross-device link

To work around this, detect the failure
and copy the file and permissions over manually.
This is not as atomic as an `os.Rename`
but it's better than complete failure.