Use clap for CLI parsing
argh has a few rough egdes,
in particular around error message customization,
and adding support for a `--version` flag.

I don't want to spend too long on trying to make that work,
so switching to clap for now.
This inflates the binary size almost 2x,
which is still less than the Go binary,
and can be addressed in a future version.