Respect core.commentChar during restacking
Git supports setting `core.commentChar` to change
which character it uses for comments in commit messages.
This affects the rebase instruction list that restack sees.

This change adds support to restack for this configuration.
If set, the specified character will be used during restacking;
both while searching for comments and when adding comments.

Resolves #72